refactor: update NavRail component and extract hook
This commit refactors the NavRail component to improve code organization and reusability. The changes include:

1. The extraction of the logic from NavRail into a custom hook called useNavRail. The hook encapsulates the state management and event handling functionality previously present in NavRail.
2. The import of the useNavRail hook in NavRail to access the extracted logic.
3. The addition of a new ThemeIconContainer component to display the ThemeIcon.
4. The import of the ThemeIcon component from the assets/images/icons/navigation folder.

These changes improve the modularity and maintainability of the NavRail component, as well as enhance the overall structure of the code.

No other files or components were affected by this refactoring.
