+++ california-dev/src/components/ui/Navigation/Desktop/NavRail.jsx
@@ -1,37 +1,23 @@
-import { useState, useEffect, useRef, useCallback } from "react";
-import { Nav, NavLinksWrapper, NavLink } from "./NavRail.styles";
+import {
+  Nav,
+  NavLinksWrapper,
+  NavLink,
+  ThemeIconContainer,
+} from "./NavRail.styles";
 import { navItemsData } from "./utils/navConfig";
 import { Logo } from "assets/images/logo";
+import { ThemeIcon } from "assets/images/icons/navigation";
+import { useNavRail } from "./hooks/useNavRail";
 
 export const NavRail = () => {
-  const [activeAnchor, setActiveAnchor] = useState(null);
-  const [pressedAnchor, setPressedAnchor] = useState(null);
-  const navLinksRefs = useRef([]);
-
-  useEffect(() => {
-    const handleGlobalMouseUp = () => {
-      setPressedAnchor(null);
-    };
-    window.addEventListener("mouseup", handleGlobalMouseUp);
-
-    return () => {
-      window.removeEventListener("mouseup", handleGlobalMouseUp);
-    };
-  }, []);
-
-  const handleMouseDown = useCallback((id) => {
-    setPressedAnchor(id);
-  }, []);
-
-  const handleMouseUp = useCallback(() => {
-    setPressedAnchor(null);
-  }, []);
-
-  const handleNavLinkClick = useCallback((id, index) => {
-    setActiveAnchor(id);
-    navLinksRefs.current[index].scrollIntoView({ behavior: "smooth" });
-  }, [navLinksRefs]);
-
+  const {
+    activeAnchor,
+    pressedAnchor,
+    navLinksRefs,
+    handleMouseDown,
+    handleMouseUp,
+    handleNavLinkClick,
+  } = useNavRail();
   return (
     <Nav>
       <Logo className="logo" />
@@ -52,6 +38,9 @@ export const NavRail = () => {
           </NavLink>
         ))}
       </NavLinksWrapper>
+      <ThemeIconContainer>
+        <ThemeIcon className="theme-icon" />
+      </ThemeIconContainer>
     </Nav>
   );
 };
